chore: Split tests (#92)
