+++ .github/workflows/ci.yaml
@@ -96,6 +96,7 @@ jobs:
       matrix:
         tox-environments:
           - integration-ha
+          - integration-replication
     name: ${{ matrix.tox-environments }}
     needs:
       - lint
