+++ tox.ini
@@ -26,6 +26,7 @@ set_env =
     tls: TEST_FILE=test_tls.py
     upgrade: TEST_FILE=test_upgrade.py
     ha: TEST_FILE=ha/test_ha.py
+    replication: TEST_FILE=ha/test_replication.py
 pass_env =
     PYTHONPATH
     CHARM_BUILD_DIR
@@ -79,7 +80,7 @@ commands =
         -m pytest -v --tb native -s {posargs} {[vars]tests_path}/unit
     poetry run coverage report
 
-[testenv:integration-{provider,password-rotation,tls,upgrade,ha}]
+[testenv:integration-{provider,password-rotation,tls,upgrade,ha,replication}]
 description = Run integration tests
 pass_env =
     {[testenv]pass_env}
